% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{BCVAR_cNIW_estimate}
\alias{BCVAR_cNIW_estimate}
\title{Estimate Conjugate Normal Inverse Wishart Bayesian Compressed VAR model}
\usage{
BCVAR_cNIW_estimate(mod_setup, keep, type, verbose = FALSE, n_chains = 1,
  n_phi = 10)
}
\arguments{
\item{keep}{Number of simulations. If set to 0, no simulation sample is created, only posterior parameters.}

\item{type}{\code{"all"} or \code{"max"}. Defines what model to takes: 
\code{"max"} takes model with best compression (lowest BIC) and \code{"all"} takes weighted average of all models.}

\item{verbose}{Whether to include messages.}

\item{n_chains}{Number of chains (for convergence). Generally not needed for conjugate BVARs. Default is 1.}

\item{n_phi}{Number of compressions for each \code{m}.}

\item{setup}{The setup provided by \code{BVAR_conj_setup}.}
}
\value{
A list with: setup parameters (priors), posterior parameters (S, Phi, Omega and v), 
and sample, if requested.
}
\description{
Obtain posterior parameters and a sample of Phi and sigma matrix
}
\details{
The function takes prior parameters (from \code{BVAR_conj_setup}) and estimates the model.
}
