% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{BCVAR_cNIW_setup}
\alias{BCVAR_cNIW_setup}
\title{Create a setup for Conjugate Normal Inverse Wishart Bayesian Compressed VAR model}
\usage{
BCVAR_cNIW_setup(series, p = 4, v_prior = NULL, Omega, S, Phi,
  include_const = TRUE)
}
\arguments{
\item{series}{Time series for estimation.}

\item{p}{Number of lags.}

\item{v_prior}{Prior value of hyperparameter nu. By default \code{m + 2}.}

\item{Omega}{Prior Omega (variance matrix for coefficient lags) matrix hyperparameter}

\item{S}{Prior covariance matrix hyperparameter}

\item{Phi}{Prior Phi (coefficient matrix) matrix hyperparameter}

\item{include_const}{Default \code{TRUE}. Whether to include the constant in the model.}
}
\value{
A list containing: Y and X matrices, X_plus and Y_plus matrices, v_prior, p and whether constant is 
included.
}
\description{
Assemble all prior and hyperparameter data from provided prior hyperparameters.
}
\details{
The function takes hyperparameter data and constructs a list of data suitable for 
\code{BCVAR_conj_est()} function. Calculations are done by a C++ function (callable to R as \code{BCVAR_conj_setup}).
}
