% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{BVAR_conj_setup}
\alias{BVAR_conj_setup}
\title{Create a setup for Conjugate Normal Inverse Wishart BVAR model}
\usage{
BVAR_conj_setup(series, lambda = c(0.2, 1, 1, 1, 100, 100), p = 1,
  delta = 1, v_prior = NULL, s2_lag = NULL, Z_mat = NULL,
  y_bar_type = c("initial", "all"), include_const = TRUE,
  carriero_hack = FALSE)
}
\arguments{
\item{series}{Time series for estimation.}

\item{lambda}{A set of lambdas (default: \code{l_1 = 0.2, l_lag = 1, l_sc = 1, l_io = 1, l_const = 100, l_exo = 100}).}

\item{p}{Number of lags.}

\item{delta}{A vector [m x 1]/ scalar / \code{"AR1"}. Used for Phi prior. If \code{delta} is set to \code{"AR1"}
then AR(1) model is used to calculate delta.}

\item{v_prior}{Prior value of hyperparameter nu. By default \code{m + 2}.}

\item{s2_lag}{Number of lags in AR() model used to estimate sig^2. By default is set to \code{p}.}

\item{Z_mat}{Matrix of exogenous variables}

\item{y_bar_type}{\code{"initial"} or \code{"all"}. Defines how y_bar is calculated for sum-of-coefficients and 
initial observation dimmy variables. Default is \code{"initial"}.}

\item{include_const}{Default \code{TRUE}. Whether to include the constant in the model.}

\item{carriero_hack}{Logical, if \code{TRUE} sigma^2 will be estimated using biased estimator
and supposed error with no square roots in dummy observations will be reproduced.}
}
\value{
A list containing: Y and X matrices, X_plus and Y_plus matrices, v_prior, p and whether constant is 
included.
}
\description{
Assemble all prior and hyperparameter data from provided lambdas.
}
\details{
The function takes hyperparameter data and constructs a list of data suitable for 
\code{BVAR_conj_estimate()} function. Calculations are done by a C++ function (callable to R as \code{bvar_setup}).
}
