\name{RcppBVAR-package}
\alias{RcppBVAR-package}
\alias{RcppBVAR}
\docType{package}
\title{
\packageTitle{RcppBVAR}
}
\description{
\packageDescription{RcppBVAR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RcppBVAR}
\packageIndices{RcppBVAR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RcppBVAR}

Maintainer: \packageMaintainer{RcppBVAR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
